# docs/api/car-pay-in-openapi.yaml
openapi: 3.0.3
info:
  title: Car Pay-in API
  version: 1.0.0
  description: |
    Car Pay-in 서비스의 전체 API 명세입니다.
    - **Car Pay-in Backend API**: AAOS 앱과 통신하는 주요 API
    - **PMS 연동 API**: 주차 관제 시스템(PMS)과의 연동 API
    - **Mock PG 연동 API**: 결제 게이트웨이(PG)와의 연동 API
servers:
  - url: https://api.carpayin.example.com
    description: Car Pay-in Backend
  - url: https://pms.example.com
    description: PMS (Parking Management System)
  - url: https://pg.example.com
    description: Mock PG (Payment Gateway)
  - url: https://card.example.com
    description: Mock Card Company
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: App Access Token (JWT)
    PmsWebhookSignature:
      type: apiKey
      in: header
      name: X-PMS-Signature
      description: PMS webhook 요청 검증용 HMAC signature

  # ─────────────────────────────────────────────
  # Reusable Schemas
  # ─────────────────────────────────────────────
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: INVALID_TOKEN
        message:
          type: string
          example: 유효하지 않은 토큰입니다.

    # ── Auth ──────────────────────────────────
...
    SimLocationRequest:
      type: object
      required: [lat, lng]
      properties:
        lat:
          type: number
          format: double
          description: 시뮬레이션 차량 위도
          example: 37.48544722
        lng:
          type: number
          format: double
          description: 시뮬레이션 차량 경도
          example: 127.03636666
        speed_kph:
          type: number
          format: double
          default: 0.0
          description: 차량 속도(km/h)
          example: 12.5
        heading:
          type: number
          format: double
          default: 0.0
          description: 진행 방향(degree)
          example: 90.0
        source:
          type: string
          default: webots
          description: 위치 업데이트 출처
          example: webots
    SimLocationResponse:
      type: object
      required: [lat, lng, speed_kph, heading, source, updated_at]
      properties:
        lat:
          type: number
          format: double
          example: 37.48544722
        lng:
          type: number
          format: double
          example: 127.03636666
        speed_kph:
          type: number
          format: double
          example: 12.5
        heading:
          type: number
          format: double
          example: 90.0
        source:
          type: string
          example: webots
        updated_at:
          type: string
          format: date-time
          description: 서버가 위치 상태를 갱신한 시각
          example: "2024-01-15T10:30:00Z"
    NavigateRequest:
      type: object
      required: [lot_id]
      properties:
        lot_id:
          type: string
          description: 길안내를 누른 제휴 주차장 ID
          example: "LOT_TEST_01"
    NavigateResponse:
      type: object
      properties:
        status:
          type: string
          enum: [registered]
          example: registered
        car_id:
          type: string
          description: 토큰에서 추출된 차량 ID
        lot_id:
          type: string
        plate:
          type: string
          description: DB에서 조회된 차량번호
    EntryWebhookRequest:
      type: object
      required: [pms_session_id, lot_id, plate, entry_time]
      properties:
        pms_session_id:
          type: string
          example: "pms_sess_001"
        lot_id:
          type: string
          example: "lot_001"
        plate:
          type: string
          example: "12가 3456"
        entry_time:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
    EntryWebhookResponse:
      type: object
      properties:
        status:
          type: string
          enum: [confirmed, not_registered]
          example: confirmed
        session_id:
          type: string
          nullable: true
        car_id:
          type: string
          nullable: true
        lot_id:
          type: string
        entry_time:
          type: string
          format: date-time
          nullable: true
...
    PmsPreRegisterRequest:
      type: object
      required: [lot_id, plate]
      properties:
        lot_id:
          type: string
          example: "lot_001"
        plate:
          type: string
          example: "12가 3456"
    PmsPreRegisterResponse:
      type: object
      properties:
        status:
          type: string
          enum: [registered]
        lot_id:
          type: string
        plate:
          type: string
    PmsLprEntryRequest:
      type: object
      required: [lot_id, plate, entry_time]
      properties:
        lot_id:
          type: string
          example: "lot_001"
        plate:
          type: string
          example: "12가 3456"
        entry_time:
          type: string
          format: date-time
    PmsLprEntryResponse:
      type: object
      properties:
        status:
          type: string
          enum: [created, existing]
        pms_session_id:
          type: string
    PmsFeeResponse:
      type: object
      properties:
        pms_session_id:
          type: string
        lot_id:
          type: string
        plate:
          type: string
        amount:
          type: integer
          example: 6000
        duration_minutes:
          type: integer
          description: 주차 시간(분)
          example: 90
        currency:
          type: string
          example: KRW
        entry_time:
          type: string
          format: date-time
        calculated_at:
          type: string
          format: date-time
    PmsPaymentCompleteRequest:
      type: object
      required: [pms_session_id, carpay_parking_session_id, carpay_tx_id, amount, currency, approval_no, idempotency_key]
      properties:
        pms_session_id:
          type: string
        carpay_parking_session_id:
          type: string
        carpay_tx_id:
          type: string
        amount:
          type: integer
          example: 6000
        currency:
          type: string
          example: KRW
        approval_no:
          type: string
        idempotency_key:
          type: string
          description: "SHA-256(session_id + car_id + amount + currency)"
    PmsPaymentCompleteResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success]
        pms_session_id:
          type: string
        carpay_tx_id:
          type: string

    # ── Mock PG ───────────────────────────────
...
paths:
...
  /sim/location:
    get:
      tags: [Parking]
      summary: 현재 시뮬레이션 차량 위치 조회
      description: |
        Webots 또는 로컬 테스트 UI가 마지막으로 저장한 시뮬레이션 차량 위치를 조회합니다.
      responses:
        '200':
          description: 현재 시뮬레이션 위치
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimLocationResponse'
    post:
      tags: [Parking]
      summary: 시뮬레이션 차량 위치 갱신
      description: |
        Webots 또는 로컬 테스트 도구가 시뮬레이션 차량 위치를 서버 메모리에 갱신합니다.
        이 endpoint는 로컬 시연/테스트 편의를 위한 상태 저장 API입니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimLocationRequest'
      responses:
        '200':
          description: 갱신된 시뮬레이션 위치
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimLocationResponse'
        '422':
          description: 요청 body validation 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /parking/navigate:
    post:
      tags: [Parking]
      summary: 제휴 주차장 길안내 → 사전 입차 알림 등록
      description: |
        사용자가 앱에서 제휴 주차장 목록의 길안내 버튼을 누를 때 호출합니다.
        lot_id만 전달하면 되며, car_id와 plate는 토큰과 DB에서 자동으로 조회합니다.

        Backend 처리:
        1. app access token 검증 및 car_id, user_id 추출
        2. DB vehicles에서 car_id로 차량과 번호판 조회
        3. vehicle_billing_keys에 active 빌링키 존재 여부 확인

        검증 완료 시:
        - carpayin-redis에 사전 입차 상태 저장 (TTL: 1시간)
          ```
          key: parking_pre_notify:{lot_id}:{plate}
          value: { user_id, car_id, lot_id, plate, status: "incoming", created_at, expires_at }
          ```
        - PMS에 차량번호 사전 등록 요청 (pms-redis에 TTL 1시간 저장)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NavigateRequest'
      responses:
        '200':
          description: 사전 입차 알림 등록 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NavigateResponse'
        '400':
          description: active 빌링키 없음 또는 차량번호 미등록
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /webhook/entry:
    post:
      tags: [Parking]
      summary: 입차 이벤트 webhook (PMS → Backend)
      description: |
        PMS LPR 카메라가 차량번호 인식 후 입차 이벤트를 이 endpoint로 전송합니다.
        
        Backend 처리:
        1. X-PMS-Signature 등으로 PMS 신뢰 요청 여부 확인
        2. Redis에서 parking_pre_notify:{lot_id}:{plate} 조회
           - 없으면: Car Pay-in 비등록 차량으로 판단, 세션 미생성
        3. Redis의 user_id, car_id를 기준으로 처리
        4. RDS parking_sessions에서 car_id의 active 세션 중복 확인
        5. Car Pay-in DB parking_sessions에 주차 세션 생성 (status=active)
        6. Redis 사전 입차 등록 상태 완료 처리
        7. MQTT/AWS IoT로 앱에 입차 확정 알림 발행
           ```
           topic: parking/confirmed/{car_id}
           payload: { session_id, car_id, lot_id, plate, entry_time, status: "active" }
           ```
        
        **멱등성**: car_id당 active 세션 1개 제약(uniq_active_session_per_car)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntryWebhookRequest'
      parameters:
        - name: X-PMS-Signature
          in: header
          required: true
          schema:
            type: string
          description: PMS webhook HMAC signature
      responses:
        '200':
          description: 입차 처리 완료. pre-notify가 없으면 status=not_registered로 반환하고 세션은 생성하지 않음.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntryWebhookResponse'
        '400':
          description: 유효하지 않은 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: PMS 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /pms/parking/pre-register:
    post:
      tags: [PMS]
      summary: 차량번호 사전 등록 (→ PMS)
      description: |
        Car Pay-in Backend가 PMS에 입차 예정 차량을 사전 등록합니다.
        PMS는 해당 차량번호와 주차장 ID를 입차 감지 대상으로 등록합니다.
        
        실제 요청 URL: `POST {PMS_URL}/parking/pre-register`
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PmsPreRegisterRequest'
      responses:
        '200':
          description: 사전 등록 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PmsPreRegisterResponse'
  /pms/lpr/entry:
    post:
      tags: [PMS]
      summary: LPR 입차 이벤트 시뮬레이션
      description: |
        Mock PMS 내부에서 LPR 입차 이벤트를 발생시켜 PMS 세션을 생성하고,
        Car Pay-in Backend의 `/webhook/entry`로 전달할 payload를 만든다.

        실제 요청 URL: `POST {PMS_URL}/lpr/entry`
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PmsLprEntryRequest'
      responses:
        '200':
          description: PMS 세션 생성 또는 기존 active 세션 반환
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PmsLprEntryResponse'
  /pms/parking/fee:
    get:
      tags: [PMS]
      summary: 주차 요금 조회 (→ PMS)
      description: |
        Car Pay-in Backend가 PMS에 현재 주차 요금을 조회합니다.
        PMS는 plate, lot_id로 active 세션을 찾아 입차 시각부터 현재까지의 요금을 계산합니다.
        
        실제 요청 URL: `GET {PMS_URL}/parking/fee`
      parameters:
        - name: lot_id
          in: query
          required: true
          schema:
            type: string
        - name: plate
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 요금 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PmsFeeResponse'
        '404':
          description: active PMS session 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /pms/payment/complete:
    post:
      tags: [PMS]
      summary: 결제 완료 통보 (→ PMS)
      description: |
        Car Pay-in Backend가 PMS에 결제 완료를 통보합니다.
        PMS는 해당 결과를 payment_requests에 저장하고, 출구 차단기 개방 기준으로 사용합니다.
        
        실제 요청 URL: `POST {PMS_URL}/payment/complete`
        
        **멱등성**: payment_requests.idempotency_key UNIQUE 제약으로 중복 저장 방지
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PmsPaymentCompleteRequest'
      responses:
        '200':
          description: 결제 완료 처리 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PmsPaymentCompleteResponse'
        '409':
          description: PMS idempotency conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ══════════════════════════════════════════
  # Mock PG 연동 API (Car Pay-in Backend → Mock PG)
  # ══════════════════════════════════════════
...